+++ wcomponents-examples/src/main/java/com/github/bordertech/wcomponents/examples/layout/FlowLayoutExample.java
@@ -1,5 +1,6 @@
 package com.github.bordertech.wcomponents.examples.layout;
 
+import com.github.bordertech.wcomponents.HeadingLevel;
 import com.github.bordertech.wcomponents.WButton;
 import com.github.bordertech.wcomponents.WHeading;
 import com.github.bordertech.wcomponents.WHorizontalRule;
@@ -22,23 +23,25 @@ public class FlowLayoutExample extends WPanel {
 	 * Creates a FlowLayoutExample.
 	 */
 	public FlowLayoutExample() {
-		add(new WHeading(WHeading.TITLE, "Flow layout examples"));
-		add(new WHorizontalRule());
+		add(new WHeading(HeadingLevel.H2, "Flow layout examples"));
+
+		// default constructor
+		WPanel flowPanel = new WPanel();
+		flowPanel.setLayout(new FlowLayout());
+		add(flowPanel);
+		addBoxes(flowPanel, 5);
 
 		// Left, no gap
-		WPanel flowPanel = new WPanel();
+		add(new WHeading(HeadingLevel.H3, "Flow layout - left, no gap"));
+		flowPanel = new WPanel();
 		flowPanel.setLayout(new FlowLayout(Alignment.LEFT));
-
-		add(new WHeading(WHeading.SECTION, "Flow layout - left, no gap"));
 		add(flowPanel);
 		addBoxes(flowPanel, 8);
 
 		// Left, with gap
+		add(new WHeading(HeadingLevel.H3, "Flow layout - left, horizontal gap"));
 		flowPanel = new WPanel();
-		flowPanel.setLayout(new FlowLayout(Alignment.LEFT, 5, 0));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - left, horizontal gap 5"));
+		flowPanel.setLayout(new FlowLayout(Alignment.LEFT, 6));
 		add(flowPanel);
 		flowPanel.add(new WButton("Button 1 text"));
 		flowPanel.add(new WButton("B2"));
@@ -46,20 +49,16 @@ public class FlowLayoutExample extends WPanel {
 		flowPanel.add(new WButton("B4"));
 
 		// Right, no gap
+		add(new WHeading(HeadingLevel.H3, "Flow layout - right, no gap"));
 		flowPanel = new WPanel();
 		flowPanel.setLayout(new FlowLayout(Alignment.RIGHT));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - right, no gap"));
 		add(flowPanel);
 		addBoxes(flowPanel, 8);
 
 		// Right, with gap
+		add(new WHeading(HeadingLevel.H3, "Flow layout - right, horizontal gap"));
 		flowPanel = new WPanel();
-		flowPanel.setLayout(new FlowLayout(Alignment.RIGHT, 5, 0));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - right, horizontal gap 5"));
+		flowPanel.setLayout(new FlowLayout(Alignment.RIGHT, 6));
 		add(flowPanel);
 		flowPanel.add(new WButton("Button 1 text"));
 		flowPanel.add(new WButton("B2"));
@@ -67,77 +66,100 @@ public class FlowLayoutExample extends WPanel {
 		flowPanel.add(new WButton("B4"));
 
 		// Vertical
+		add(new WHeading(HeadingLevel.H3, "Flow layout - vertical"));
 		flowPanel = new WPanel();
 		flowPanel.setLayout(new FlowLayout(Alignment.VERTICAL));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - vertical"));
 		add(flowPanel);
 		addBoxes(flowPanel, 5);
 
 		// Center
+		add(new WHeading(HeadingLevel.H3, "Flow layout - center, horizontal gap"));
 		flowPanel = new WPanel();
-		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 5, 0));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - center, horizontal gap 5"));
+		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 6));
 		add(flowPanel);
 		addBoxes(flowPanel, 5);
 		flowPanel.add(new WText("And some more text"));
 
 		// Mixed (Outer center, center inner)
+
+		add(new WHeading(HeadingLevel.H3, "Flow layout - center + vertical, vertical gap"));
 		flowPanel = new WPanel();
 		flowPanel.setLayout(new FlowLayout(Alignment.CENTER));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - center + vertical, vertical gap 5"));
 		add(flowPanel);
 
 		WPanel inner = new WPanel();
 		flowPanel.add(inner);
-		inner.setLayout(new FlowLayout(Alignment.VERTICAL, 0, 5));
+		inner.setLayout(new FlowLayout(Alignment.VERTICAL, 6));
 		addBoxes(inner, 5);
 
 		// Content Alignment - TOP
+		add(new WHeading(HeadingLevel.H3, "Flow layout - center, horizontal gap, TOP"));
 		flowPanel = new WPanel();
-		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 5, 0, ContentAlignment.TOP));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - center, horizontal gap 5, TOP"));
+		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 6, ContentAlignment.TOP));
 		add(flowPanel);
 		addBoxesWithDiffContent(flowPanel, 5);
 		flowPanel.add(new WText("some text"));
 
 		// Content Alignment - MIDDLE
+		add(new WHeading(HeadingLevel.H3, "Flow layout - center, horizontal gap, MIDDLE"));
 		flowPanel = new WPanel();
-		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 5, 0, ContentAlignment.MIDDLE));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - center, horizontal gap 5, MIDDLE"));
+		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 6, ContentAlignment.MIDDLE));
 		add(flowPanel);
 		addBoxesWithDiffContent(flowPanel, 5);
 		flowPanel.add(new WText("some text"));
 
 		// Content Alignment - BASELINE
+		add(new WHeading(HeadingLevel.H3, "Flow layout - center, horizontal gap, BASELINE"));
 		flowPanel = new WPanel();
-		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 5, 0, ContentAlignment.BASELINE));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - center, horizontal gap 5, BASELINE"));
+		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 6, ContentAlignment.BASELINE));
 		add(flowPanel);
 		addBoxesWithDiffContent(flowPanel, 5);
 		flowPanel.add(new WText("some text"));
 
 		// Content Alignment - BOTTOM
+		add(new WHeading(HeadingLevel.H3, "Flow layout - center, horizontal gap, BOTTOM"));
 		flowPanel = new WPanel();
-		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 5, 0, ContentAlignment.BOTTOM));
-
-		add(new WHorizontalRule());
-		add(new WHeading(WHeading.SECTION, "Flow layout - center, horizontal gap 5, BOTTOM"));
+		flowPanel.setLayout(new FlowLayout(Alignment.CENTER, 6, ContentAlignment.BOTTOM));
 		add(flowPanel);
 		addBoxesWithDiffContent(flowPanel, 5);
 		flowPanel.add(new WText("some text"));
 
+		// deprecated constructors
+		add(new WHorizontalRule());
+		add(new WHeading(HeadingLevel.H2, "Checks of deprecated constructors"));
+		// These are here to check we have not broken the old constructors.
+
+		add(new WHeading(HeadingLevel.H3, "Flow layout - left, horizontal gap and ignored vertical gap"));
+		flowPanel = new WPanel();
+		flowPanel.setLayout(new FlowLayout(Alignment.LEFT, 6, 12));
+		add(flowPanel);
+		addBoxes(flowPanel, 12);
+		flowPanel.add(new WText("some text"));
+		flowPanel.add(new WText("some more text"));
+
+		add(new WHeading(HeadingLevel.H3, "Vertical, ignored horizontal gap and implemented vertical gap"));
+		flowPanel = new WPanel();
+		flowPanel.setLayout(new FlowLayout(Alignment.VERTICAL, 6, 12));
+		add(flowPanel);
+		addBoxes(flowPanel, 5);
+		flowPanel.add(new WText("some text"));
+
+
+		add(new WHeading(HeadingLevel.H3, "Left, horizontal gap and ignored vertical gap, content align bottom"));
+		flowPanel = new WPanel();
+		flowPanel.setLayout(new FlowLayout(Alignment.LEFT, 6, 12, ContentAlignment.BOTTOM));
+		add(flowPanel);
+		addBoxesWithDiffContent(flowPanel, 5);
+		flowPanel.add(new WText("some text"));
+		flowPanel.add(new WText("some more text"));
+
+
+		add(new WHeading(HeadingLevel.H3, "VERTICAL, ignored horizontal gap, vertical gap, ignored content align bottom"));
+		flowPanel = new WPanel();
+		flowPanel.setLayout(new FlowLayout(Alignment.VERTICAL, 6, 12, ContentAlignment.BOTTOM));
+		add(flowPanel);
+		addBoxes(flowPanel, 5);
+		flowPanel.add(new WText("some text"));
 	}
 
 	/**
@@ -163,7 +185,7 @@ public class FlowLayoutExample extends WPanel {
 	private static void addBoxesWithDiffContent(final WPanel panel, final int amount) {
 		for (int i = 1; i <= amount; i++) {
 			WPanel content = new WPanel(WPanel.Type.BOX);
-			content.setLayout(new FlowLayout(FlowLayout.VERTICAL, 0, 3));
+			content.setLayout(new FlowLayout(FlowLayout.VERTICAL, 3));
 			for (int j = 1; j <= i; j++) {
 				content.add(new WText(Integer.toString(i)));
 			}